+++ .eslintignore
@@ -1,1 +1,2 @@
 flow-typed/
+node_modules/
