+++ package.json
@@ -85,6 +85,7 @@
     "webpack": "^1.13.1",
     "webpack-dev-server": "^1.16.2",
     "webpack-manifest-plugin": "^1.1.0",
+    "webpack-md5-hash": "^0.0.5",
     "yargs": "^6.6.0"
   }
 }
