+++ src/lib/dom/index.js
@@ -29,3 +29,11 @@ export function addScript (src: string) {
   script.src = src;
   window.document.head.appendChild(script);
 }
+
+export function addStyleSheet (src: string) {
+  const style = document.createElement('link');
+  style.href = src;
+  style.setAttribute('rel', 'stylesheet');
+  style.setAttribute('type', 'text/css');
+  document.head && document.head.appendChild(style);
+}
